libft

+++ libft/ft_bzero.c
@@ -6,11 +6,13 @@
 /*   By: tvan-gij <tvan-gij@student.19.be>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/01/17 13:11:11 by tvan-gij          #+#    #+#             */
-/*   Updated: 2021/01/17 13:11:25 by tvan-gij         ###   ########.fr       */
+/*   Updated: 2021/01/22 16:02:45 by tvan-gij         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-void	ft_bzero(void *s, unsigned int n)
+#include <stdlib.h>
+
+void	ft_bzero(void *s, size_t n)
 {
 	char *str;
