+++ libft/ft_memccpy.c
@@ -6,16 +6,18 @@
 /*   By: tvan-gij <tvan-gij@student.19.be>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/01/17 13:14:16 by tvan-gij          #+#    #+#             */
-/*   Updated: 2021/01/18 18:41:49 by tvan-gij         ###   ########.fr       */
+/*   Updated: 2021/01/22 16:06:20 by tvan-gij         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-void	*ft_memccpy(void *dst, const void *src, int c, unsigned int n)
+#include <stdlib.h>
+
+void	*ft_memccpy(void *dst, const void *src, int c, size_t n)
 {
 	unsigned char	occurs;
 	char			*dest;
 	const char		*source;
-	unsigned int	i;
+	size_t			i;
 
 	occurs = c;
 	dest = dst;
