+++ libft/ft_memchr.c
@@ -6,15 +6,17 @@
 /*   By: tvan-gij <tvan-gij@student.19.be>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/01/17 13:14:50 by tvan-gij          #+#    #+#             */
-/*   Updated: 2021/01/17 13:15:03 by tvan-gij         ###   ########.fr       */
+/*   Updated: 2021/01/22 16:09:05 by tvan-gij         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-void	*ft_memchr(const void *s, int c, unsigned int n)
+#include <stdlib.h>
+
+void	*ft_memchr(const void *s, int c, size_t n)
 {
 	unsigned char	occurs;
 	unsigned char	*str;
-	unsigned int	i;
+	size_t			i;
 
 	occurs = c;
 	str = (unsigned char *)s;
