+++ libft/ft_memcmp.c
@@ -6,11 +6,13 @@
 /*   By: tvan-gij <tvan-gij@student.19.be>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/01/17 13:15:15 by tvan-gij          #+#    #+#             */
-/*   Updated: 2021/01/17 13:15:28 by tvan-gij         ###   ########.fr       */
+/*   Updated: 2021/01/22 16:10:08 by tvan-gij         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-int		ft_memcmp(const void *s1, const void *s2, unsigned int n)
+#include <stdlib.h>
+
+int		ft_memcmp(const void *s1, const void *s2, size_t n)
 {
 	unsigned char	*str1;
 	unsigned char	*str2;
