+++ libft/ft_memcpy.c
@@ -6,11 +6,13 @@
 /*   By: tvan-gij <tvan-gij@student.19.be>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/01/17 13:15:44 by tvan-gij          #+#    #+#             */
-/*   Updated: 2021/01/18 18:42:16 by tvan-gij         ###   ########.fr       */
+/*   Updated: 2021/01/22 16:04:25 by tvan-gij         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-void	*ft_memcpy(void *dst, const void *src, unsigned int n)
+#include <stdlib.h>
+
+void	*ft_memcpy(void *dst, const void *src, size_t n)
 {
 	char		*dest;
 	const char	*source;
