+++ libft/ft_memmove.c
@@ -6,11 +6,13 @@
 /*   By: tvan-gij <tvan-gij@student.19.be>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/01/17 13:16:15 by tvan-gij          #+#    #+#             */
-/*   Updated: 2021/01/17 13:16:34 by tvan-gij         ###   ########.fr       */
+/*   Updated: 2021/01/22 16:07:28 by tvan-gij         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-void	*ft_memmove(void *dst, const void *src, unsigned int len)
+#include <stdlib.h>
+
+void	*ft_memmove(void *dst, const void *src, size_t len)
 {
 	char		*dest;
 	const char	*source;
