+++ libft/ft_strlcat.c
@@ -6,13 +6,15 @@
 /*   By: tvan-gij <tvan-gij@student.19.be>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/01/17 13:22:01 by tvan-gij          #+#    #+#             */
-/*   Updated: 2021/01/18 18:42:40 by tvan-gij         ###   ########.fr       */
+/*   Updated: 2021/01/22 16:16:49 by tvan-gij         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-unsigned int	ft_strlenc(const char *s)
+#include <stdlib.h>
+
+size_t	ft_strlenc(const char *s)
 {
-	unsigned int	i;
+	size_t	i;
 
 	i = 0;
 	while (s[i] != '\0')
@@ -20,11 +22,11 @@ unsigned int	ft_strlenc(const char *s)
 	return (i);
 }
 
-unsigned int	ft_strlcat(char *dst, const char *src, unsigned int dstsize)
+size_t	ft_strlcat(char *dst, const char *src, size_t dstsize)
 {
-	unsigned int	i;
-	unsigned int	j;
-	unsigned int	lendst;
+	size_t	i;
+	size_t	j;
+	size_t	lendst;
 
 	i = 0;
 	j = ft_strlenc(src);
