+++ libft/ft_strlcpy.c
@@ -6,15 +6,17 @@
 /*   By: tvan-gij <tvan-gij@student.19.be>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/01/17 13:22:17 by tvan-gij          #+#    #+#             */
-/*   Updated: 2021/01/18 18:43:06 by tvan-gij         ###   ########.fr       */
+/*   Updated: 2021/01/22 16:13:53 by tvan-gij         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-unsigned int	ft_strlcpy(char *dst, const char *src, unsigned int dstsize)
+#include <stdlib.h>
+
+size_t	ft_strlcpy(char *dst, const char *src, size_t dstsize)
 {
-	unsigned int	i;
-	char			*dest;
-	const char		*source;
+	size_t		i;
+	char		*dest;
+	const char	*source;
 
 	i = 0;
 	dest = dst;
