+++ libft/ft_strnstr.c
@@ -6,15 +6,17 @@
 /*   By: tvan-gij <tvan-gij@student.19.be>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/01/17 13:23:51 by tvan-gij          #+#    #+#             */
-/*   Updated: 2021/01/17 13:23:59 by tvan-gij         ###   ########.fr       */
+/*   Updated: 2021/01/22 16:21:00 by tvan-gij         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-char	*ft_strnstr(const char *haystack, const char *needle, unsigned int len)
+#include <stdlib.h>
+
+char	*ft_strnstr(const char *haystack, const char *needle, size_t len)
 {
-	unsigned int	i;
-	unsigned int	j;
-	char			*ans;
+	size_t	i;
+	size_t	j;
+	char	*ans;
 
 	i = 0;
 	ans = (char *)haystack;
